requirements.txt
Flask>=3.0,<4.0
Flask-Login>=0.6,<0.7
Flask-SQLAlchemy>=3.1,<4.0
Authlib>=1.3,<2.0
python-dotenv>=1.0,<2.0
requests>=2.31,<3.0
google-auth>=2.29,<3.0
google-api-python-client>=2.120,<3.0

# PostgreSQL (OBRIGATÓRIO EM PRODUÇÃO)
psycopg2-binary>=2.9,<3.0

# Servidor WSGI (produção Linux)
gunicorn>=21.2,<22.0
